Move add_controlled_RY_gate
