+++ packages/chem/quri_parts/chem/utils/excitations.py
@@ -14,10 +14,10 @@ from typing_extensions import TypeAlias
 
 from quri_parts.circuit import (
     LinearMappedUnboundParametricQuantumCircuit,
-    LinearParameterFunction,
     Parameter,
     ParameterOrLinearFunction,
 )
+from quri_parts.circuit.utils.controlled_rotations import add_controlled_RY_gate
 
 #: Alias of ``tuple[int, int]`` which represents the set of orbital indices involved in
 #: excitation. The first element is the index of an occupied orbital and the second
@@ -77,7 +77,7 @@ def add_single_excitation_circuit(
         p_fn = {param: 0.5 * val for param, val in param_fn.items()}
 
     circuit.add_CNOT_gate(*excitation_indices)
-    _add_controlled_RY_gate(circuit, excitation_indices[1], excitation_indices[0], p_fn)
+    add_controlled_RY_gate(circuit, excitation_indices[1], excitation_indices[0], p_fn)
     circuit.add_CNOT_gate(*excitation_indices)
     return circuit
 
@@ -145,20 +145,5 @@ def add_double_excitation_circuit(
     circuit.add_H_gate(excitation_indices[3])
     circuit.add_CNOT_gate(excitation_indices[0], excitation_indices[2])
     circuit.add_CNOT_gate(excitation_indices[2], excitation_indices[3])
-
-    return circuit
-
-
-def _add_controlled_RY_gate(
-    circuit: LinearMappedUnboundParametricQuantumCircuit,
-    control_index: int,
-    target_index: int,
-    param_fn: LinearParameterFunction,
-) -> LinearMappedUnboundParametricQuantumCircuit:
-    inv_sign_param_fn = {param: -1.0 * val for param, val in param_fn.items()}
-    circuit.add_ParametricRY_gate(target_index, param_fn)
-    circuit.add_CNOT_gate(control_index, target_index)
-    circuit.add_ParametricRY_gate(target_index, inv_sign_param_fn)
-    circuit.add_CNOT_gate(control_index, target_index)
 
     return circuit
